+++ gcc/rust/ast/rust-ast-collector.h
